+++ ServerData.cs
@@ -158,5 +158,5 @@ public record RouteInfo(string Route, string Method, bool IsLegacy = false);
 
 public static class APIVersion
 {
-    public const string Version = "1.9.6";
+    public const string Version = "1.9.7";
 }